Avoid "conflicting properties" warning with bound Window.visibility
Since d0644b040eb09e1ed147ff15e7c926c9e318cbca we emit a warning
"Conflicting properties 'visible' and 'visibility'" if you set or bind
both the Window.visible and Window.visibility property in a conflicting
way. But it was just noise to emit this warning if only `visibility` is
being set. `visible` is false by default; but if you make the window
visible by setting visibility, it should be ok. So we follow the pattern
of having an extra bool to store whether it was explicitly set.

Added a test to verify behavior with 4 combinations: visible unset, true
or false, and visibility set or unset. This demonstrates that if there
really is a conflict, `visibility` "wins"; and if there is no conflict,
we now suppress the warning.

Fixes: QTBUG-35244
Change-Id: I4c8c75d2d84e903c94545d0d3fb61ea5e4269ae5
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
(cherry picked from commit 1022dd338e49bfc297f94a39383a2043cbb4eb62)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>