QML: Register all builtins
We need run time access to the QQmlType instances of all the builtins in
order to properly coerce types without special casing everything all the
time.

Since we can now have QML types without metaobjects, we need to check
for the metaobject in a few places where we didn't need to check before.

Change-Id: Ib22cbb12c60ebdce4897c3f3338851e8b925926f
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>