Doc: Fix linking and read-only status for Window.visibility properties
QDoc has problems linking to a topic that has identical name to the
topic the link appears in, as is the case when linking from 'visibility'
property to 'visibility' attached property.

Add a unique \keyword for the former and use it for linking.

Explicitly set the attached property as \readonly.

Pick-to: 6.5 6.2
Task-number: QTBUG-113015
Change-Id: Ia0eb99fd2436565c697ca23629134b691acecc1b
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>