Fix SplitView orientation bug
When SplitView have no handle, and orientation changed, then in this
case positioning does not work correctly. For example, if we have
horizontal handle with size (width x 10), then after orientation changed
to vertical, the handle will take the (width x height) instead of (10 x
height). In this patch, we just swap the width and height when the
orientation changes.

Fixes: QTBUG-88195
Pick-to: 6.5
Change-Id: I75d946f5574e2e5af6f76e4ba5191d9704f15428
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>