Doc: remove 'readonly' mark from Control::focusReason
The property is declared as read-write, and it makes sense for it to
stay that way. For example, due to a bug in ListView key navigation,
client code might want to adjust focusReason on newly focused delegates.
Or maybe you are developing a dual-input(touch+keys) device, so you want
to change focusReason to Tab/BackTab whenever any key is pressed after
using touch screen, so that a larger styled focus ring would appear.

However, clarify that Control::focusReason typically shouldn't need to
be set by the user. Also mark its replacement as an action item for
Qt 7 as per the 6.7 API review discussions.

Task-number: QTBUG-117596
Pick-to: 6.7 6.8
Change-Id: Ifef3c3e76ef3d5be1131b2527a04fd8a29f054f6
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>