qmlls: support find type definition for script identifiers
Add QQmlJSScope to JS identifiers in the QQmlJSScope and resolve them in
the QQmlJSScope resolving step.
Use them to find the type definition of script identifiers.
Test the new feature.

Todo for after FF: find out why only the part after the
"component <name>:" is found (which should be enough for real usage) for
JS identifiers. In an ideal world, it would find the "<name">" and just
highlight it.

Fixes: QTBUG-111410
Change-Id: Id83ec08fd4c66a9c7468e3bec19ba09f4ce8634f
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>