qmlls: definition support for scriptexpressions
Add "go-to-definition"-support in qmlls for scriptexpressions.
Make sure that the QQmlJSLogger used to construct the Dom contains the
correct FileName, such that the QQmlJSScope also contain the filename.
This allows to find the definitions of scriptexpressions using the
QQmlJSScopes.

Add findDefinition-support to QmlLSUtils to find the definitions from a
given DomItem, and the boilerplate in qqmlgotodefinitionsupport for
qmlls to be able to use the method in QQmlLSUtils.

Extract some reused code into methods to avoid code-duplication.

Also add some tests.

Task-number: QTBUG-111409
Change-Id: I80d2345b00f02c9a7008afbb01f9b7020a259e09
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>