config.json: List state names alphabetically
We need to follow a specific pattern for the generation of the names,
especially when dealing with multiple combined states. Follow an
alphabetic pattern in this case, where the states are ordered
alphabetically in the name. This makes things more consistent and in
this way, we can just list the state names alphabetically in the
"currentStates" property in the QML templates.
Also add all the missing states to the config file.

Task-number: QTBUG-111560
Change-Id: Ic5b0500b8ec0ea01519b032686d829979f4f3c78
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>