Override default ApplicationWindow for native windows style
The ApplicationWindow quick controls with native windows style uses
dark palettes by default even though we dont't support it (from Qt6.5).
But when other quick controls such as ComboBox, if used within
ApplicationWindow, it shows up with light palette. This is because
windows style overrides controls such as ComboBox with their native
style and this inturn updates the palettes to light through its plugin.

To fix it, this patchset adds explicit ApplicationWindow quick control
for windows style and it overrides default with native style imported
within it.

Fixes: QTBUG-113469
Pick-to: 6.5
Change-Id: Ia0eb72966e05bafa7a06427db8adcb41dc30ed2d
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>