qmlls: Fix TextSynchronization
The offset from a given range was incorrectly calculated in
findBlockByNumber because textblocks that are used to obtain that
position were mistakenly ignoring the newline character. This were
leading us to setting a different text than what we actually typed.
Add an extra text block if the text ends with \n.

Add a manual test but exclude it from tst_qmlformat as it is not a valid
qml document.

Fixes: QTBUG-113725
Change-Id: Ifb1b4121fd8af46df7627d986303ae67e170e50c
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>