QtQml: Move inline component registry to QQmlMetaTypeData
QQmlTypePrivate is supposed to be immutable. "Amending" an existing
QQmlTypePrivate with inline components after the fact is risky. It's
hard to guarantee that no other thread accesses the data in between.

Also, this way we need one lookup table for all the inline components
rather than a separate (mostly empty) one for each type.

Change-Id: I7a06c0faca1f8244dd2bdc7aa775abd8e2f92125
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>