Sanitize QSG_RHI_HDR handling
Add an 'auto' mode which just means use some HDR
format if supported, use SDR otherwise. This format will be
scRGB, but that's not exposed in the API. This is more realistic
for a typical application (think a video player f.ex.).

Then expose this as a usual QQuickGraphicsConfiguration flag, but
mark the flag API as \internal for now, because the full story
for HDR support in Qt Quick is not quite clear. However, some
applications, that know what they are doing, may want to take
this into use already.

Change-Id: I3fc79246528cc89cf27d13bb016189d509498225
Reviewed-by: Andy Nichols <andy.nichols@qt.io>