Expose isPropertySet of Layout attached data
The Layout attached data stores whether a property was set or not.
This is useful to know from outside the Layout attached property
and this patch exposes this information. It was not stored for two
properties (alignment, margins), which is also added with this patch.

This is also done in preparation for the LayoutProxyItem.

Change-Id: Ia1ec33fa1f1f5836857ceba7999d7a0e750f1a66
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>