StyleGenerator: ensure we don't resolve a non existing state in SpinBox
If you hover an indicator of a SpinBox and then move the
mouse quickly outside the control, a warning message
related to the config will be printed. The reason
is that the control will emit changes to control.hovered
before changes to up.hovered (and down.hovered), which
means that state name of the config will briefly
resolve to an unexisting state. A simple solution is
to also check that the control itself is being
hovered, and not just the indicators.

Change-Id: I99e097557c949381543f3182682a4dfa5c7be2c5
Reviewed-by: Doris Verria <doris.verria@qt.io>