Stylegenerator: Add workaround for fully rounded corners
BorderImage scales horizontally/vertically by stretching the middle
area between the corners. In case the corners are fully rounded, ie.:
half of the control's size, then there is no middle area to stretch,
so the BorderImage can't stretch the image in that direction and
will just draw nothing in the middle.
That is why we need to make sure that in case the left/right and
top/bottom corners are the same and are half the image size, we have
at least a 1px middle area to stretch.

We also need to make sure that the exported offsets from the
corner radii are no bigger than half the size of the rectangle,
as bigger values are not meaningful. It just means that the
corners are overlapping.

Change-Id: Ia30accbb90b64a293ca2927138db062257f0974c
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>