QObjectWrapper: Use metaTypeFromJS for call arguments
We want the precise type there, not some approximation that happens to
be a QVariant. This exposes that we had a few conversions that were
possible with QVariant::convert() but not with our builtin type
coercion.

Change-Id: I44c57a22bad8268de3d4398721e1c63b18009dfc
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit 67d0c0821276b57eb8ec57ecad7b2e3f06e50667)
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>