internal doc: Add scenario Two touchpoints dragging two DragHandlers
This adds to the "ideal" scenarios being described on the
Qt Quick Pointer event delivery page.

This time we try using PlantUML to generate whole UML Communication
Diagrams rather than post-editing in Inkscape.
- build and install umlquick, which now generates PlantUML output by
  default
- `$ qml ~/dev/umlquick/examples/objectInstancesTimer.qml
   ../../snippets/pointerHandlers/pinchAndDragHandlers.qml`
- close the qml window after the .puml file is generated
- open the puml file in a text editor, and begin adding labels
  corresponding to messages on existing inter-object connections.
  Add new dotted-line connections with labels as necessary.
- preview as necessary:
  `plantuml -tpng pinchAndDragHandlers-doublePressDelivery.puml`
Doxygen will run plantuml on its own when building the docs later on.

One downside is that as with most graph-layout algorithms, adding extra
edges between nodes can completely restructure the layout; so it's not
possible to make successive iterations of the .puml file with the same
layout of object instances, with only the extra edges being different
between them. But so far we only have a single diagram in this scenario,
for the press event, and explain the subsequent move event in prose.

Change-Id: I5c41d7e7d6f5e2d2c27e024789e56e6cc7159c8b
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>