QQuickLabsPlatformMenuItem: accept KeySequence in shortcut
The documentation claims that the shortcut of a menuitem can be bound
to a QKeySequence (or at least the qml version of it) but menuitem
actually only supports int or strings (that are wrapped into a
QVariant).

Implement the missing codepath for passing QKeySequence in the QVariant
for C++ users, instead of constructing a garbage QKeySequence from the
string obtained by QKeySequence::toString().

Fixes: QTBUG-113743
Change-Id: I6f4c659a79a4c23de34be92770815d9e14a6a554
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>
(cherry picked from commit a4de7389bf90ab62c5d12391f9bb0d03da9b4b82)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>