Software renderer: Add floating point nodes to dirty list
m_obscuredRegion avoids overpainting of nodes that are not updated.
However, due to the integer precision of QRegion, this mechanism
does not work for nodes with a bounding rectangle with floats.
Currently we add boundingRectMin to m_obscuredRegion, however,
which is not sufficient. Nodes where the boundingRectMin is
different from the real boundingRect (and boundingRectMax) can not
be added correctly to m_obscuredRegion. Therefore these nodes are
now also set to dirty and updated.

Fixes: QTBUG-113745
Change-Id: I5cc5540ed45593b09b312a1704459e95bebab521
Reviewed-by: Andy Nichols <andy.nichols@qt.io>