Reset touchPoint position after it was remaped
The position of touch points is modified in the function updateTouchData
when triggered from a filter. The position is never reset and thus wrong
in other parts of the code. This patch resets the position after the
required stuff in the offset coordinate system is done.

Fixes: QTBUG-105862
Change-Id: Idd12af0d509e46d0ac74fb920c70e9ec6411c255
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit 99a48bef8468ce6155d9898479a01062988624d3)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>