QML: Do not leak memory if QQmlData is manipulated from ctor
This creates another QQmlData object on the heap, which is rather
wasteful. However, we cannot really avoid it since it's all too easy to
trigger an operation like this.

If it happens, use the QQmlData it has created, and ignore the memory
we've allocated inline. While we're at it, make the memory ownership a
required argument to the ctor.

Fixes: QTBUG-114186
Change-Id: I0568768c9ec13c94db79bb162c9eeb76f75f2a55
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit a3389d6bf238c46816fd1133c1258102e36c4b10)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>