Match fontEngine glyph cache key to lifespan of owner
The glyph cache key is based on the RHI instance. We clean up the font
cache when the QSGRenderContext is deleted.

For a regular QQuickWindow this is effectively the same. For a
QQuickWidget the RHI instance has a lifespan of the parent widget, but
the render context changes.

Task-number: QTBUG-113426
Task-number: QTBUG-112306
Pick-to: 6.6 6.5
Change-Id: Ifd901ace6caaff9a08352901956865f5f58e06ef
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>