StyleGenerator: only create images for atoms that are visible in the design
As it stood, we would only skip creating images for atoms
that themselves where marked as hidden by Figma. But an
atom can be logically hidden if a parent is hidden. So
make the logic more robust, and give the designer more
freedom to choose which child in the design should be
hidden in order to hide the equivalent image in QQC.

Change-Id: I2a33c87fff0dfbc6317c02f56c6e1ef20b069845
Reviewed-by: Doris Verria <doris.verria@qt.io>