qmllint: Detect unqualified id access in StoreNameSloppy
In QQmlJSTypePropagator::generate_StoreName, we would return early in a
few error cases (after calling setError). That is okay for the compiler,
which cares about setError, but qmllint is not interested in compiler
errors by default.
This would result in us not issuing warnings about unqualified id
accesses.
Fix this by also logging the errors via the logger.

Fixes: QTBUG-114467
Pick-to: 6.5 6.6
Change-Id: I10ad10e5c93762b3d9d1fb50d3cabcff2c0b7a4f
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>