MouseArea: don't ignore double-click events
In 72651a50f83aa72998822312c7b5c6235d28978f
QQuickItem::mouseDoubleClickEvent() started to ignore double-clicks by
default, which is consistent with the fact that it ignores the other
pointer event types. But now we have to worry about subclasses that
override mouseDoubleClickEvent() and call the base class implementation.

d7fac6923a6d4e4ac7dc22458256366968acbdb3 tried to fix it but neglected
the case when the MouseArea does not have an onDoubleClicked signal
handling script. Since the QQuickMouseEvent object that we emit to QML
won't be accepted if isDoubleClickConnected() is false, and since the
code before 72651a50f83aa72998822312c7b5c6235d28978f was leaving the
event accepted regardless of whether QQuickMouseEvent was accepted, we
should not need to check it now either. Perhaps we should care about the
case when onDoubleClicked sets accepted to false, but so far that
doesn't seem very useful either: if you accept the press, a parent
MouseArea will not see either the press or the double-click; if you
ignore the press, you won't see the double-click, and a parent MouseArea
will see both by default. tst_QQuickMouseArea::clickThrough() tests
accepted = false in onDoubleClicked but not onPressed, and only with
mouse, not touch.

Added tst_QQuickMouseArea::doubleTap(); also moved the autotest from
d7fac6923a6d4e4ac7dc22458256366968acbdb3 which has nothing to do with
pointer handlers.

Fixes: QTBUG-112434
Fixes: QTBUG-109393
Pick-to: 6.6 6.5 6.5.2
Change-Id: I426827c20cdb2373e77744987dffba59cd941edc
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Doris Verria <doris.verria@qt.io>