AbstractButton: don't ignore double-click events
After 72651a50f83aa72998822312c7b5c6235d28978f
QQuickItem::mouseDoubleClickEvent() calls ignore(). That probably
causes a behavior change in any subclass that overrides and then
unconditionally calls the base class implementation.

QQuickTextEdit::mouseDoubleClickEvent() has it like this:

    if (!event->isAccepted())
        QQuickImplicitSizeItem::mouseDoubleClickEvent(event);

But here there is no reason for the event not to be accepted: it
arrives pre-accepted, and Button does not give the user the opportunity
to ignore it (like MouseArea does). And QQuickControl does not override
mouseDoubleClickEvent() either, so the line we're removing would call
QQuickItem::mouseDoubleClickEvent() which we don't want.

Pick-to: 6.5 6.6
Change-Id: I2e3f72ed6aededbfaeb1a3174a6c9edc6efd1d5c
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>