CMake: Remove superfluous depfile in QML type registration
In _qt_internal_qml_type_registration we created a depfile that
contained the *_metatypes.json files of target's SOURCES.

This depfile was used in the custom command that produces
${target}_qmltyperegistration.cpp.

It doesn't seem necessary to create a depfile, and one could use the
DEPENDS option of add_custom_command instead. And indeed, commit
ce950d619aef3ad0534544b2e029a6f025b256a5 added the same dependencies
this way. Presumably, this was done to fix dependencies for older CMake
versions and generators that do not support depfiles.

This commit removes the creation of the depfile.

This double dependency declaration was found when investigating
QTBUG-106683, and it turns out that removing the depfile fixes the
problem of re-generating ${target}_qmltyperegistration.cpp for Visual
Studio generators.

Pick-to: 6.2 6.5 6.6
Task-number: QTBUG-106683
Change-Id: I6cf909dfb4a9b31b76599b7ad35ad6f595ca0891
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>