Move textselection example to manual test
Apparently it's always been incomplete, a relic from Nokia: there's no
means of entering the "menu" state, and nothing to detect a long-press
as the text says you can do.

On mobile platforms, the platform input method provides selection
handles; on embedded platforms, Qt Virtual Keyboard does it. So this is
generally unnecessary, but perhaps it could be revamped into a
demonstration of an alternative way to add text-selection handles.

Pick-to: 6.6 6.5
Change-Id: I231db14e5fbb4e82e8d732153e3d3843a080b409
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>