QML models: Make most properties FINAL to prevent accidental shadowing
qmlsc won't be able to efficiently compile JS expressions containing
non-FINAL properties, since it doesn't know if the property will be
shadowed during runtime.

Shadowing properties are also an anti-pattern at best, and something we
strongly encourage users to not do.

[ChangeLog][Important Behavior Changes][QmlModels] Most properties for
types in the QtQml.Models module are now FINAL, meaning that they can no
longer be shadowed by declaring new properties with the same names.
A warning will be printed out to the console, when a FINAL property is
shadowed. We recommend that users rename those properties to avoid
potential unexpected behavior changes.

Task-number: QTBUG-108739
Pick-to: 6.6
Change-Id: Ief8b4b19272996afbc989d355ddc99e44427ea72
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>