Don't include qcoregraphics_p.h on non-apple platforms
The qcoregraphics_p.h header is only included by the build system if
qtbase is built for an apple platform.

This means that people that want to use the iOS style on
non-apple platforms (which isn't officially supported), have to
comment out the line that includes this header.

There is no downside to using #ifdefs here, and it makes it more
convenient for non-apple users to hack our ways to into using the iOS
style.

Change-Id: I2ddd1cc2bfacea303da14d4477de7952af836142
Reviewed-by: Doris Verria <doris.verria@qt.io>