QmlCachegen: Avoid adding a block's jumpOrigin a second time
Each code block is reached either by a jump or simply by a fallthrough
when the next instruction is executed. To model code flow, the origins
of jumps (explicit or falltrough) to each block are recorded.

The fallthrough jumpOrigins were added in two places: processJump and
endInstruction. In some situations both would add the origin leading to
duplication.

Instead, have endInstruction be the only place where origins are added.
And only have processJump add the block to m_basicBlocks to ensure it
can be found by endInstruction.

Change-Id: Ic5e4d7ab7c80f1f97f8373db9bf4b300f0f0ebd1
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>