StyleGenerator: support both light and dark mode
This change will add support for producing images
and other assets for both light and dark mode.

The way it's implemented, is that you now specify which
themes that you want to generate in the config.json file.
Each control will then be exported one time for each theme
listed, where the variable ${Theme} will be set to the name
of the theme being processed. This can then be used to alter
the figma paths to point to different components per theme.
E.g, the first time ButtonTemplate${Theme} is processed, it will
be translated to ButtonTemplateLight, and the second time it will
be ButtonTemplateDark.

The exported images/assets will be placed inside respective theme
folders, together with the config.json file that describes them.

In order to support this, the QQuickFigmaConfigReader has
been modified to pick up the current theme from the system in
order to resolved the correct config.json file. When the
theme changes, the controlsChanged signal will be emitted, which
will cause all QML bindings to update as well. The result is
that all images will change.

Change-Id: I8f7e3f74b3ea699165b327a0776342af3074685e
Reviewed-by: Doris Verria <doris.verria@qt.io>